Create LICENSE
feat: change go mod

build(deps): bump golang.org/x/net

Bumps [golang.org/x/net](https://github.com/golang/net) from 0.0.0-20211029224645-99673261e6eb to 0.7.0.
- [Release notes](https://github.com/golang/net/releases)
- [Commits](https://github.com/golang/net/commits/v0.7.0)

---
updated-dependencies:
- dependency-name: golang.org/x/net
  dependency-type: indirect
...


feat: add proxy

Update sdk.go

Update sdk.go

fix: Temperature type

fix: type

+++ audio.go
@@ -21,7 +21,7 @@ type RequestTranscription struct {
 
 	// number Optional Defaults to 0
 	// The sampling temperature, between 0 and 1. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic. If set to 0, the model will use log probability to automatically increase the temperature until certain thresholds are hit.
-	Temperature int `json:"temperature,omitempty"`
+	Temperature float64 `json:"temperature,omitempty"`
 
 	// string Optional
 	// The language of the input audio. Supplying the input language in ISO-639-1 format will improve accuracy and latency.
@@ -61,7 +61,7 @@ type RequestTranslation struct {
 
 	// number Optional Defaults to 0
 	// The sampling temperature, between 0 and 1. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic. If set to 0, the model will use log probability to automatically increase the temperature until certain thresholds are hit.
-	Temperature int `json:"temperature,omitempty"`
+	Temperature float64 `json:"temperature,omitempty"`
 }
 
 type ResponseTranslation struct {
