+++ completions.go
@@ -31,13 +31,13 @@ type RequestCompletions struct {
 	// What sampling temperature to use, between 0 and 2. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic.
 	//
 	// We generally recommend altering this or top_p but not both.
-	Temperature int `json:"temperature,omitempty"`
+	Temperature float64 `json:"temperature,omitempty"`
 
 	// number Optional Defaults to 1
 	// An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered.
 	//
 	// We generally recommend altering this or temperature but not both.
-	TopP int `json:"top_p,omitempty"`
+	TopP float64 `json:"top_p,omitempty"`
 
 	// integer Optional Defaults to 1
 	// How many completions to generate for each prompt.
@@ -67,13 +67,13 @@ type RequestCompletions struct {
 	// Number between -2.0 and 2.0. Positive values penalize new tokens based on whether they appear in the text so far, increasing the model's likelihood to talk about new topics.
 	//
 	// See more information about frequency and presence penalties. https://platform.openai.com/docs/api-reference/parameter-details
-	PresencePenalty int `json:"presence_penalty,omitempty"`
+	PresencePenalty float64 `json:"presence_penalty,omitempty"`
 
 	// number Optional Defaults to 0
 	// Number between -2.0 and 2.0. Positive values penalize new tokens based on their existing frequency in the text so far, decreasing the model's likelihood to repeat the same line verbatim.
 	//
 	// See more information about frequency and presence penalties. https://platform.openai.com/docs/api-reference/parameter-details
-	FrequencyPenalty int `json:"frequency_penalty,omitempty"`
+	FrequencyPenalty float64 `json:"frequency_penalty,omitempty"`
 
 	// integer Optional Defaults to 1
 	// Generates best_of completions server-side and returns the "best" (the one with the highest log probability per token). Results cannot be streamed.
