+++ edits.go
@@ -22,13 +22,13 @@ type RequestEdits struct {
 	// What sampling temperature to use, between 0 and 2. Higher values like 0.8 will make the output more random, while lower values like 0.2 will make it more focused and deterministic.
 	//
 	// We generally recommend altering this or top_p but not both.
-	Temperature int `json:"temperature,omitempty"`
+	Temperature float64 `json:"temperature,omitempty"`
 
 	// number Optional Defaults to 1
 	// An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered.
 	//
 	// We generally recommend altering this or temperature but not both.
-	TopP int `json:"top_p,omitempty"`
+	TopP float64 `json:"top_p,omitempty"`
 }
 
 type ResponseEdits struct {
